LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/sticky/view-timeline-sticky-offscreen-2-expected.txt

Subject

FAIL View timeline bottom-sticky during entry and top-sticky during exit. assert_equals: Effect at the start of the active phase: cover 0% to cover 100% expected "0.3" but got "0.366667"
